+++ flor-de-lotus/flor-de-lotus/src/main/java/flor_de_lotus/paciente/dto/ViewTotalPacientesPorAno.java
@@ -1,13 +1,13 @@
 package flor_de_lotus.paciente.dto;
 
 public class ViewTotalPacientesPorAno {
-    private Integer qtd;
+    private Long qtd;
 
-    public ViewTotalPacientesPorAno(Integer totalPacientesAtivos) {
+    public ViewTotalPacientesPorAno(Long totalPacientesAtivos) {
         this.qtd = totalPacientesAtivos;
     }
 
-    public Integer getQtd() {
+    public Long getQtd() {
         return qtd;
     }
 }
